requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
motor==3.5.1
pydantic[email]==2.9.1
python-dotenv==1.0.1
supabase==2.7.2
openai==1.46.0
razorpay==1.4.2
httpx==0.27.2
